Put the onboarding view in the splashVC
